backend/requirements.txt
# Django Core
Django>=4.2,<5.0
djangorestframework>=3.14
djangorestframework-simplejwt>=5.3
django-cors-headers>=4.3

# ASGI & WebSocket
daphne>=4.0
channels>=4.0
channels-redis>=4.1

# Database
psycopg2-binary>=2.9  # PostgreSQL (production)

# HTTP Requests
requests>=2.31

# Image handling
Pillow>=10.0
